+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Non associative operators
 ### Changed
 - One function `get` for operators in API
+- `make_appl` does not use the table of operators
 
 ## [1.0.1] -- 2021-01-16
 ### Fixed